Immediately call the callback
This is very not correct, but affinity hangs on startup otherwise.
I tried calling the unix space function, but the callback is in the
pe address space likely. And I'm not sure how to translate the calls
from opencl -> wine server -> the pe process. I tried a wrapper function
but it still crashes.